requirements.txt
robotframework==6.1
robotframework-seleniumlibrary==6.1.1
selenium==4.17.2
robotframework
openpyxl
